Deduplicate style code
